requirements.txt
keras
autokeras
tensorflow
sqlalchemy
numpy
Pillow
flask
keras
flask_restplus
flask-sqlalchemy
flask-migrate
flask_restful_swagger
requests
psycopg2-binary
flask-cors
pyjwt